S12P21B105-181 FEAT: 소셜 로그인 기능 추가

See merge request s12-ai-speech-sub1/S12P21B105!60

+++ server/speechfy/src/main/java/com/ssafy/speechfy/service/SongService.java
@@ -96,7 +96,11 @@ public class SongService {
      * 완성곡 삭제
      */
     @Transactional
-    public void deleteSongById(int songId) {
+    public void deleteSongById(int songId, int userId) {
+        Song song = songRepository.findById(songId);
+        if (song.getUser().getId() != userId) {
+            return;
+        }
         songRepository.deleteById(songId);
     }
